# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "MBD - @e.bollo Service",
    "termsOfService": "https://www.pagopa.gov.it/",
    "title": "EBollo 2.0 - Service for partner",
    "version": "1.2.11-2-PAGOPA-3747-ebollo-2"
  },
  "servers": [
    {
      "url": "",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/info": {
      "get": {
        "description": "Return OK if application is started",
        "operationId": "healthCheck_2",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppInfo"
                }
              }
            },
            "description": "OK",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ],
        "summary": "health check",
        "tags": [
          "Home"
        ]
      },
      "parameters": [
        {
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "in": "header",
          "name": "X-Request-Id",
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/v1/info": {
      "get": {
        "description": "Return OK if application is started",
        "operationId": "healthCheck_1",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppInfo"
                }
              }
            },
            "description": "OK",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ],
        "summary": "health check",
        "tags": [
          "MBD - v1",
          "Home"
        ]
      },
      "parameters": [
        {
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "in": "header",
          "name": "X-Request-Id",
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/v1/organizations/{organization-fiscal-code}/mbd": {
      "parameters": [
        {
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "in": "header",
          "name": "X-Request-Id",
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "deprecated": true,
        "description": "Return mbd data for payment on requirement",
        "operationId": "getMdb",
        "parameters": [
          {
            "description": "Organization fiscal code",
            "in": "path",
            "name": "organization-fiscal-code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetMbdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCartResponse"
                }
              }
            },
            "description": "OK",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": 400,
                  "title": "Bad Request",
                  "detail": "<detail.message>"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCartErrorResponse"
                }
              }
            },
            "description": "Service unavailable",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ],
        "summary": "getMbd",
        "tags": [
          "MBD - v1"
        ]
      }
    },
    "/v1/organizations/{organization-fiscal-code}/receipt/{nav}": {
      "get": {
        "deprecated": true,
        "description": "Return the Marca da Bollo Digitale",
        "operationId": "getPaymentReceipts",
        "parameters": [
          {
            "description": "Organization fiscal code",
            "in": "path",
            "name": "organization-fiscal-code",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Notice number",
            "in": "path",
            "name": "nav",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMdbReceipt"
                }
              }
            },
            "description": "OK",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "title": "Not Found",
                  "detail": "<detail.message>"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "status": 500,
                  "title": "Internal Server Error",
                  "detail": "<detail.message>"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            },
            "description": "Service unavailable",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ],
        "summary": "getPaymentReceipt",
        "tags": [
          "MBD - v1"
        ]
      },
      "parameters": [
        {
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "in": "header",
          "name": "X-Request-Id",
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/v2/info": {
      "get": {
        "description": "Return OK if application is started",
        "operationId": "healthCheck",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AppInfo"
                }
              }
            },
            "description": "OK",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ],
        "summary": "health check",
        "tags": [
          "MBD - v2",
          "Home"
        ]
      },
      "parameters": [
        {
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "in": "header",
          "name": "X-Request-Id",
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/v2/organizations/{organization-fiscal-code}/mbd": {
      "parameters": [
        {
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "in": "header",
          "name": "X-Request-Id",
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "description": "Return mbd data for payment on requirement ",
        "operationId": "getMdbV2",
        "parameters": [
          {
            "description": "Organization fiscal code",
            "in": "path",
            "name": "organization-fiscal-code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetMbdRequestV2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCartResponse"
                }
              }
            },
            "description": "OK",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "example": {
                  "status": 400,
                  "title": "Bad Request",
                  "detail": "<detail.message>"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCartErrorResponse"
                }
              }
            },
            "description": "Service unavailable",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ],
        "summary": "getMbdV2",
        "tags": [
          "MBD - v2"
        ]
      }
    },
    "/v2/organizations/{organization-fiscal-code}/noticeNumbers/{nav}/mbd": {
      "get": {
        "description": "Return the Marca da Bollo Digitale",
        "operationId": "getPaymentReceiptsV2",
        "parameters": [
          {
            "description": "Organization fiscal code",
            "in": "path",
            "name": "organization-fiscal-code",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Notice number",
            "in": "path",
            "name": "nav",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMdbReceipt"
                }
              }
            },
            "description": "OK",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            },
            "description": "Bad Request",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "status": 404,
                  "title": "Not Found",
                  "detail": "<detail.message>"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            },
            "description": "Not Found",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "example": {
                  "status": 500,
                  "title": "Internal Server Error",
                  "detail": "<detail.message>"
                },
                "schema": {
                  "$ref": "#/components/schemas/ProblemJson"
                }
              }
            },
            "description": "Service unavailable",
            "headers": {
              "X-Request-Id": {
                "description": "This header identifies the call",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          }
        ],
        "summary": "getPaymentReceiptV2",
        "tags": [
          "MBD - v2"
        ]
      },
      "parameters": [
        {
          "description": "This header identifies the call, if not passed it is self-generated. This ID is returned in the response.",
          "in": "header",
          "name": "X-Request-Id",
          "schema": {
            "type": "string"
          }
        }
      ]
    }
  },
  "components": {
    "schemas": {
      "AppInfo": {
        "type": "object",
        "properties": {
          "environment": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "GetCartErrorResponse": {
        "type": "object",
        "properties": {
          "errorUrl": {
            "type": "string",
            "description": "The error URL provided in the request.",
            "example": "https://url.it"
          }
        }
      },
      "GetCartResponse": {
        "type": "object",
        "properties": {
          "checkoutRedirectUrl": {
            "type": "string",
            "description": "Redirect URL to Checkout page.",
            "example": "https://dev.checkout.pagopa.it.it/c/123"
          },
          "mbdDownloadLink": {
            "type": "string",
            "description": "URL to download MBD receipt.",
            "example": "https://api.platform.pagopa.it/mbd/mbd-payment/v1/00000000000/receipt/300996299640641032"
          },
          "nav": {
            "type": "string",
            "description": "Notice number, used to download MBD receipt.",
            "example": "300996299640641032"
          }
        }
      },
      "GetMbdRequest": {
        "required": [
          "idCIService",
          "returnUrls"
        ],
        "type": "object",
        "properties": {
          "idCIService": {
            "type": "string",
            "description": "Identifier of the association between the organization and MDB payment service.",
            "example": "04"
          },
          "paymentNotices": {
            "maxItems": 1,
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentNotice"
            }
          },
          "returnUrls": {
            "$ref": "#/components/schemas/ReturnUrls"
          }
        }
      },
      "GetMbdRequestV2": {
        "required": [
          "idCIService",
          "returnUrls"
        ],
        "type": "object",
        "properties": {
          "idCIService": {
            "type": "string",
            "description": "Identifier of the association between the organization and MDB payment service.",
            "example": "04"
          },
          "paymentNotices": {
            "maxItems": 1,
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentNoticeV2"
            }
          },
          "returnUrls": {
            "$ref": "#/components/schemas/ReturnUrlsV2"
          }
        }
      },
      "GetMdbReceipt": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "description": "The MBD receipt.",
            "items": {
              "type": "string",
              "description": "The MBD receipt.",
              "format": "byte"
            }
          }
        }
      },
      "PaymentNotice": {
        "required": [
          "amount",
          "documentHash",
          "email",
          "firstName",
          "fiscalCode",
          "lastName",
          "province"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "type": "integer",
            "description": "MBD amount.",
            "format": "int64",
            "example": 16
          },
          "documentHash": {
            "maxLength": 44,
            "minLength": 44,
            "type": "string",
            "description": "Document hash."
          },
          "email": {
            "type": "string",
            "description": "Debtor mail.",
            "example": "mario.rossi@test.test"
          },
          "firstName": {
            "type": "string",
            "description": "Debtor name.",
            "example": "Mario"
          },
          "fiscalCode": {
            "pattern": "^(?:[A-Z]{6}\\d{2}[A-Z]\\d{2}[A-Z]\\d{3}[A-Z]|\\d{11})$",
            "type": "string",
            "description": "Debtor fiscal code.",
            "example": "00000000000000000"
          },
          "lastName": {
            "type": "string",
            "description": "Debtor surname.",
            "example": "Rossi"
          },
          "province": {
            "type": "string",
            "description": "Debtor province.",
            "example": "RM"
          }
        }
      },
...
      "ProblemJson": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string",
            "description": "A human readable explanation specific to this occurrence of the problem.",
            "example": "There was an error processing the request"
          },
          "status": {
            "maximum": 600,
            "minimum": 100,
            "type": "integer",
            "description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
            "format": "int32",
            "example": 400
          },
          "title": {
            "type": "string",
            "description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "The API key to access this function app.",
        "in": "header",
        "name": "Ocp-Apim-Subscription-Key",
        "type": "apiKey"
      }
    }
  }
}